cricbuzz_mysql/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0
plotly>=5.18.0
python-dotenv>=1.0.0
mysql-connector-python>=8.3.0
